Added a Basic State Machine for the Drug State and a Pickup Drug Item for Testing
The only effect I've got it doing right now is to increase/decrease movement speed, but should be simple enough to pass in a VFX or something else to add more effects.
The timings and severity of the withdrawals feels quite off at the moment but should be an easy rework.